Version 5.1.0,  Revision 20101,  Hash 62f948d06
MODIFIED *  .gitignore Makefile config/configure.ac config/m4/yambo_specific.m4 config/mk/actions/clean.mk config/mk/actions/compile_external_libraries.mk config/mk/actions/compile_internal_libraries.mk config/mk/actions/compile_yambo.mk config/mk/actions/compile_yambo_libraries.mk config/mk/actions/dependencies.mk config/mk/defs.mk.in config/mk/functions/cleaning.mk config/mk/functions/mk_external_yambo_lib.mk config/mk/functions/mk_internal_lib.mk config/mk/functions/mk_src.mk config/mk/libraries.mk config/mk/targets.mk config/version/version.m4 configure sbin/compilation/dependencies.sh sbin/compilation/helper.sh sbin/compilation/mk/functions.mk sbin/compilation/mk/make.inc.in sbin/compilation/mk/makefile sbin/compilation/mk/rules.mk sbin/compilation/mk/static_variables.mk

NEW *  sbin/compilation/configure_generated_files.sh sbin/compilation/special_sources.sh

DELETED *  config/mk/variables.mk

RENAMED *  config/stamps/.empty -> config/stamps_and_lists/.empty

Changes:
- More compilation related coding

Patch sent by:  Andrea Marini <andrea.marini@cnr.it>

+++ sbin/compilation/mk/functions.mk
@@ -5,14 +5,14 @@
 # Driver
 #--------
 define driver
- $(PREFIX)(eval $(cc) $(cflags) $(precomp_flags) $(linclude) -L$(libdir) -D_$@ -c $(libdir)/yambo/driver/src/driver/driver.c $(LOGID)> /dev/null)
+ $(PREFIX)(eval $(cc) $(cflags) $(precomp_flags) $(lf90include) -L$(libdir) -D_$@ -c $(libdir)/yambo/driver/src/driver/driver.c $(LOGID)> /dev/null)
  $(msg)
 endef
 #
 # Linking
 #---------
 define link
- $(PREFIX)(eval $(fc) $(fcflags) $(lf90include) $(lf90libinclude) -o $@ driver.o $(objs) $(libs) ) > /dev/null
+ $(PREFIX)(eval $(fc) $(fcflags) $(lf90include) $(lf90libinclude) -o $@ driver.o $(objs) $(libs) $(LOGID)> /dev/null)
  $(msg)
 endef
 #
@@ -20,7 +20,7 @@ endef
 #-------------
 define c_elemental_compilation
  $(rm_command)
- $(PREFIX)(eval $(cc) $(cflags) $(precomp_flags) $(linclude) -c $(srcdir)/$(wdir)/$*.c $(LOGID) > /dev/null)
+ $(PREFIX)(eval $(cc) $(cflags) $(precomp_flags) $(lf90include) -c $(srcdir)/$(wdir)/$*.c $(LOGID)> /dev/null)
  $(msg)
 endef
 define f77_elemental_compilation
@@ -48,42 +48,42 @@ define F90_local_elemental_compilation
  $(msg)
 endef
 define F90_elemental_compilation
+ if ! test -f $(compdir)/stamps_and_lists/$(target).stamp; 
+ then 
  $(rm_command)
  $(PREFIX)(eval $(fpp) $(precomp_flags) $(lf90include) $(lf90libinclude) $(srcdir)/$(wdir)/$*.F > $*.tmp_source)
  $(PREFIX)($(srcdir)/sbin/replacer.sh $*.tmp_source)
  $(PREFIX)(mv $*.tmp_source_space $*$(f90suffix))
  $(PREFIX)($(fc) -c $(fcflags) $(lf90include) $(lf90libinclude) $*$(f90suffix) $(LOGID)> /dev/null)
  $(msg)
+ fi
 endef
 #
 # Object(s) operations
 #----------------------
-define o_save
- $(PREFIX)(if test "$(keep_objs)" = "yes"; \
- then $(compdir)/sbin/objects_store.sh $(objects_lock); rm -f $(moduledep_file) $(modlist_file); fi)
- $(PREFIX)(if test "$(keep_objs)" = "no"; then rm -f $(moduledep_file) $(modlist_file); fi)
-endef
-define o_and_mod_clean
- $(PREFIX)(if test ! -f $(objects_lock) && test "$(keep_objs)" = "no" ; then \
- find . \( -name '*.o' -o -name '*.mod' -o -name '__*' \) | xargs rm -f ; \
- touch $(objects_lock); rm -f $(moduledep_file) $(modlist_file); fi)
-endef
-define mk_mod_dir
- $(PREFIX)(if test ! -d $(modinclude); then echo "creating folder $(modinclude)" ; fi)
- $(PREFIX)(if test ! -d $(modinclude); then mkdir $(modinclude) ; fi)
-endef
+#define o_save
+# $(PREFIX)(if test "$(keep_objs)" = "yes"; \
+# then $(compdir)/sbin/objects_store.sh $(objects_lock); rm -f $(moduledep_file) $(modlist_file); fi)
+# $(PREFIX)(if test "$(keep_objs)" = "no"; then rm -f $(moduledep_file) $(modlist_file); fi)
+#endef
+#define o_and_mod_clean
+# $(PREFIX)(if test ! -f $(objects_lock) && test "$(keep_objs)" = "no" ; then \
+# find . \( -name '*.o' -o -name '*.mod' -o -name '__*' \) | xargs rm -f ; \
+# touch $(objects_lock); rm -f $(moduledep_file) $(modlist_file); fi)
+#endef
+#define mk_mod_dir
+# $(PREFIX)(if test ! -d $(modinclude); then echo "creating folder $(modinclude)" ; fi)
+# $(PREFIX)(if test ! -d $(modinclude); then mkdir $(modinclude) ; fi)
+#endef
 define modmove
- $(PREFIX)(test `$(mfiles) | wc -l` -eq 0 || $(mfiles) > $(modlist_file))
- $(PREFIX)(test `$(mfiles) | wc -l` -eq 0 ||  mv *.mod $(modinclude))
-endef
-define mk_moduledep_file
- $(PREFIX)(cd $(srcdir)/$(wdir); $(srcdir)/sbin/moduledep.sh $(objs) > $(compdir)/$(wdir)/$(moduledep_file))
+ $(PREFIX)MODS=`find . -name '*.mod'`;for modfile in $$MODS ; do mv $$modfile $(compdir)/include; done
 endef
 define mk_lib
- $(PREFIX)(eval $(ar) $(arflags) $(target) $(objs) > /dev/null)
+ $(PREFIX)(eval $(ar) $(arflags) $(target) $(objs) $(LOGID)> /dev/null)
  $(PREFIX)(mv $(target) $(libdir))
  $(PREFIX)(chmod u+x $(libdir)/$(target))
- $(PREFIX)(echo "\t $(target)" )
+ $(PREFIX)(echo "\t [$(wdir)] $(target)" )
+ $(PREFIX)(touch $(compdir)/config/stamps_and_lists/$(target).stamp)
 endef
 #
 # Utils
@@ -92,6 +92,6 @@ define dircheck
  $(PREFIX)if test ! -d $(exec_prefix); then mkdir $(exec_prefix);fi
 endef
 define msg
- $(PREFIX)(echo "\t [$(target)] $*" )
+ $(PREFIX)(echo "\t [$(wdir)] $*" )
 endef
